auth-service/requirements.txt
Django==4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
redis
django-cors-headers